+++ sys/net/pipex.h
@@ -217,7 +217,7 @@ struct mbuf           *pipex_l2tp_input (struct mbuf *, int off, struct pipex_se
 struct pipex_session  *pipex_l2tp_userland_lookup_session_ipv4 (struct mbuf *, struct in_addr);
 struct pipex_session  *pipex_l2tp_userland_lookup_session_ipv6 (struct mbuf *, struct in6_addr);
 struct mbuf           *pipex_l2tp_userland_output (struct mbuf *, struct pipex_session *);
-int                   pipex_ioctl (struct pipex_iface_context *, int, caddr_t);
+int                   pipex_ioctl (struct pipex_iface_context *, u_long, caddr_t);
 void                  pipex_session_init_mppe_recv(struct pipex_session *, int,
 int, u_char *);
 void                  pipex_session_init_mppe_send(struct pipex_session *, int,
